Spring Security를 이용한 username&password 인증 로직 구현

+++ src/main/java/com/realworld/study/post/application/PostService.java
@@ -1,6 +1,12 @@
 package com.realworld.study.post.application;
 
+import com.realworld.study.auth.util.UsernamePasswordAuthUtils;
+import com.realworld.study.exception.domain.IsNotAuthorThisPostException;
+import com.realworld.study.exception.domain.MemberNotFoundException;
+import com.realworld.study.exception.domain.PostNotFoundException;
+import com.realworld.study.member.domain.Email;
 import com.realworld.study.post.domain.Post;
+import com.realworld.study.post.domain.PostQueryRepository;
 import com.realworld.study.post.domain.PostRepository;
 import com.realworld.study.post.presentation.dto.PostCreateRequest;
 import com.realworld.study.post.presentation.dto.PostUpdateRequest;
@@ -11,6 +17,7 @@ import com.realworld.study.member.domain.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,50 +26,67 @@ import org.springframework.transaction.annotation.Transactional;
 @Service  //TODO CGLIB (PROXY) 상속
 public class PostService {
     private final PostRepository postRepository;
+    private final PostQueryRepository postQueryRepository;
     private final MemberRepository memberRepository;
 
-    public PostResponse createPost(final PostCreateRequest postCreateRequest, final Member member) {
-        //TODO 원래는 인증 객체에서 member 식별자를 꺼내서, memberRepository에서 멤버를 조회해서 사용해야 한다.
-        // 현재는 인증을 구현하지 않아 아래의 방식으로 대체한다.
-        memberRepository.save(member);
+    public PostResponse createPost(final PostCreateRequest postCreateRequest,
+            final Authentication authentication) {
+        Member member = findMemberBy(authentication);
 
-        Post post = getPostBy(postCreateRequest, member);
+        Post post = createPostBy(postCreateRequest, member);
         postRepository.save(post);
 
         return PostResponse.from(post);
     }
 
-    private Post getPostBy(final PostCreateRequest dto, final Member member) {
+    private Post createPostBy(final PostCreateRequest dto, final Member member) {
         return new Post(dto.getTitle(), dto.getContents(), member);
     }
 
-    public PostResponse updatePost(final Long postId, final PostUpdateRequest postUpdateRequest) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("없는 게시글 입니다."));
+    public PostResponse updatePost(final Long postId, final PostUpdateRequest postUpdateRequest,
+            final Authentication authentication) {
+        Post post = findPostBy(postId);
+        validateAuthor(post, authentication);
 
-        post.update(postUpdateRequest.getTitle(), postUpdateRequest.getContents());  //TODO 이 부분의 DTO 의존성을 없앨 수는 없을까?
+        post.update(postUpdateRequest.getTitle(), postUpdateRequest.getContents());
+
         return PostResponse.from(post);
     }
 
-    public PostDeleteResponse deletePost(final Long postId) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("없는 게시글 입니다."));
+    public PostDeleteResponse deletePost(final Long postId, final Authentication authentication) {
+        Post post = findPostBy(postId);
+        validateAuthor(post, authentication);
 
         postRepository.delete(post);
         return new PostDeleteResponse(true);
     }
 
     @Transactional(readOnly = true)
-    public PostResponse getPost(Long postId) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("없는 게시글 입니다."));
-
+    public PostResponse getPost(final Long postId) {
+        Post post = findPostBy(postId);
         return PostResponse.from(post);
     }
 
     @Transactional(readOnly = true)
-    public Page<PostResponse> getPosts(Pageable pageable) {
-        Page<Post> posts = postRepository.pagedPosts(pageable);
-        return posts.map(PostResponse::from);
+    public Page<PostResponse> getPosts(final Pageable pageable) {
+        return postQueryRepository.pagedPosts(pageable);
+    }
+
+    private Post findPostBy(final Long postId) {
+        return postRepository.findById(postId)
+                .orElseThrow(PostNotFoundException::new);
+    }
+
+    private void validateAuthor(final Post post, final Authentication authentication) {
+        Member member = findMemberBy(authentication);
+        if (!member.isAuthorOf(post)) {
+            throw new IsNotAuthorThisPostException();
+        }
+    }
+
+    private Member findMemberBy(final Authentication authentication) {
+        String email = UsernamePasswordAuthUtils.getEmail(authentication);
+        return memberRepository.findByEmail(new Email(email))
+                .orElseThrow(MemberNotFoundException::new);
     }
 }
